+++ smartech_app/lib/app_inbox/widgets/smt_simple_notification_view.dart
@@ -1,7 +1,6 @@
 import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_html/flutter_html.dart';
 import 'package:smartech_app/app_inbox/utils/utils.dart';
 import 'package:smartech_app/deep_link_screen.dart';
 import 'package:smartech_app/navigator.dart';